uc: Use uint8_t for all port_ids.
256 ports ought to be enough for anyone.
As the port name is a char, only 256 can have unique names anyway.

+++ uc/hal_arduino/hal_arduino.h
@@ -18,18 +18,18 @@
 
 typedef struct TF_Port {
     //external
-    int chip_select_pin;
+    uint8_t chip_select_pin;
     char port_name;
 } TF_Port;
 
 typedef struct TF_HalContext {
     SPISettings spi_settings;
     TF_Port *ports;
-    size_t port_count;
+    uint8_t port_count;
     TF_HalCommon hal_common;
 } TF_HalContext;
 
-int tf_hal_create(struct TF_HalContext *ctx, TF_Port *ports, size_t port_count) TF_ATTRIBUTE_NONNULL_ALL;
+int tf_hal_create(struct TF_HalContext *ctx, TF_Port *ports, uint8_t port_count) TF_ATTRIBUTE_NONNULL_ALL;
 int tf_hal_destroy(TF_HalContext *hal) TF_ATTRIBUTE_NONNULL_ALL;
 
 #endif
