+++ uc/hal_arduino_esp32/hal_arduino_esp32.h
@@ -17,7 +17,7 @@
 #include <SPI.h>
 
 typedef struct TF_Port {
-    int chip_select_pin;
+    uint8_t chip_select_pin;
     uint8_t spi;
     char port_name;
 } TF_Port;
@@ -27,11 +27,11 @@ typedef struct TF_HalContext {
     SPIClass hspi;
     SPIClass vspi;
     TF_Port *ports;
-    size_t port_count;
+    uint8_t port_count;
     TF_HalCommon hal_common;
 } TF_HalContext;
 
-int tf_hal_create(struct TF_HalContext *hal, TF_Port *ports, size_t port_count) TF_ATTRIBUTE_NONNULL_ALL;
+int tf_hal_create(struct TF_HalContext *hal, TF_Port *ports, uint8_t port_count) TF_ATTRIBUTE_NONNULL_ALL;
 int tf_hal_destroy(TF_HalContext *hal) TF_ATTRIBUTE_NONNULL_ALL;
 
 #endif
